Make sort mixin work on other properties
`SortByNameMixin` is working well for classes where we want to sort
objects by name. But it’s limited to the `name` attribute.

This commit makes it more generic, so that subclass-ees can choose
which attribute to sort on.

This would also let us implement enhancements ascending/descending
sorting in the future, for example.