Fix upload button's accessible name if errors
When an upload returned errors, we were adding a
label to the button to associate them with it.
This assumed they would be added to the existing
accessible name, derived from the button text, but
it instead replaced it.

This adds the original name to the start of the
label, for non-visual users.